backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
torch>=2.0.0
torchaudio>=2.0.0
whisperx>=3.1.0
openai-whisper>=20231117
